backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.0.1

# OpenAI Integration
openai==1.3.7
python-dotenv==1.0.0

# Optional: For better performance
# tensorflow-cpu==2.15.0  # Use this instead of tensorflow if you don't have GPU